refactor: logInlineCompletionSessionResultsHandler
