+++ server/aws-lsp-codewhisperer/src/language-server/inline-completion/codeWhispererServer.ts
@@ -1,65 +1,32 @@
 import {
     CancellationToken,
     InitializeParams,
-    InlineCompletionItemWithReferences,
     InlineCompletionListWithReferences,
-    InlineCompletionTriggerKind,
     InlineCompletionWithReferencesParams,
     LogInlineCompletionSessionResultsParams,
-    Position,
-    Range,
     Server,
-    TextDocument,
-    ResponseError,
-    LSPErrorCodes,
-    WorkspaceFolder,
 } from '@aws/language-server-runtimes/server-interface'
-import { autoTrigger, getAutoTriggerType, getNormalizeOsName, triggerType } from './auto-trigger/autoTrigger'
-import {
-    CodeWhispererServiceToken,
-    GenerateSuggestionsRequest,
-    GenerateSuggestionsResponse,
-    Suggestion,
-    SuggestionType,
-} from '../../shared/codeWhispererService'
-import { CodewhispererLanguage, getRuntimeLanguage, getSupportedLanguageId } from '../../shared/languageDetection'
-import { mergeEditSuggestionsWithFileContext, truncateOverlapWithRightContext } from './mergeRightUtils'
-import { CodeWhispererSession, SessionManager } from './session/sessionManager'
+import { getSupportedLanguageId } from '../../shared/languageDetection'
+import { SessionManager } from './session/sessionManager'
 import { CodePercentageTracker } from './codePercentage'
-import { getCompletionType, getEndPositionForAcceptedSuggestion, getErrorMessage, safeGet } from '../../shared/utils'
-import { getIdeCategory, makeUserContextObject } from '../../shared/telemetryUtils'
-import { fetchSupplementalContext } from '../../shared/supplementalContextUtil/supplementalContextUtil'
-import { textUtils } from '@aws/lsp-core'
+import { safeGet } from '../../shared/utils'
+import { makeUserContextObject } from '../../shared/telemetryUtils'
 import { TelemetryService } from '../../shared/telemetry/telemetryService'
 import { AcceptedInlineSuggestionEntry, CodeDiffTracker } from './codeDiffTracker'
-import {
-    AmazonQError,
-    AmazonQServiceConnectionExpiredError,
-    AmazonQServiceInitializationError,
-} from '../../shared/amazonQServiceManager/errors'
+import { AmazonQServiceInitializationError } from '../../shared/amazonQServiceManager/errors'
 import { AmazonQBaseServiceManager } from '../../shared/amazonQServiceManager/BaseAmazonQServiceManager'
 import { getOrThrowBaseTokenServiceManager } from '../../shared/amazonQServiceManager/AmazonQTokenServiceManager'
 import { AmazonQWorkspaceConfig } from '../../shared/amazonQServiceManager/configurationUtils'
-import { hasConnectionExpired } from '../../shared/utils'
 import { getOrThrowBaseIAMServiceManager } from '../../shared/amazonQServiceManager/AmazonQIAMServiceManager'
-import { WorkspaceFolderManager } from '../workspaceContext/workspaceFolderManager'
 import path = require('path')
-import { getRelativePath } from '../workspaceContext/util'
+
 import { UserWrittenCodeTracker } from '../../shared/userWrittenCodeTracker'
 import { RecentEditTracker, RecentEditTrackerDefaultConfig } from './tracker/codeEditTracker'
 import { CursorTracker } from './tracker/cursorTracker'
 import { RejectedEditTracker, DEFAULT_REJECTED_EDIT_TRACKER_CONFIG } from './tracker/rejectedEditTracker'
-import { getAddedAndDeletedChars } from './diffUtils'
-import {
-    emitPerceivedLatencyTelemetry,
-    emitServiceInvocationFailure,
-    emitServiceInvocationTelemetry,
-    emitUserTriggerDecisionTelemetry,
-} from './telemetry'
 import { CodeWhispererController } from './codewhispererController'
-const { editPredictionAutoTrigger } = require('./auto-trigger/editPredictionAutoTrigger')
+import { DocumentChangedListener } from './documentChangedListener'
 
-const EMPTY_RESULT = { sessionId: '', items: [] }
 export const FILE_URI_CHARS_LIMIT = 1024
 export const FILENAME_CHARS_LIMIT = 1024
 export const CONTEXT_CHARACTERS_LIMIT = 10240
@@ -67,9 +34,6 @@ export const CONTEXT_CHARACTERS_LIMIT = 10240
 export const CodewhispererServerFactory =
     (serviceManager: () => AmazonQBaseServiceManager): Server =>
     ({ credentialsProvider, lsp, workspace, telemetry, logging, runtime, sdkInitializator }) => {
-        let lastUserModificationTime: number
-        let timeSinceLastUserModification: number = 0
-
         const sessionManager = SessionManager.getInstance()
 
         // AmazonQTokenServiceManager and TelemetryService are initialized in `onInitialized` handler to make sure Language Server connection is started
@@ -93,151 +57,8 @@ export const CodewhispererServerFactory =
         const rejectedEditTracker = RejectedEditTracker.getInstance(logging, DEFAULT_REJECTED_EDIT_TRACKER_CONFIG)
         let editsEnabled = false
 
-        let apiController: CodeWhispererController
-
-        const onInlineCompletionHandler = async (
-            params: InlineCompletionWithReferencesParams,
-            token: CancellationToken
-        ): Promise<InlineCompletionListWithReferences> => {
-            return apiController.onInlineCompletion(params, token)
-        }
-
-        // Schedule tracker for UserModification Telemetry event
-        const enqueueCodeDiffEntry = (
-            session: CodeWhispererSession,
-            acceptedSuggestion: Suggestion,
-            addedCharactersForEdit?: string
-        ) => {
-            const endPosition = getEndPositionForAcceptedSuggestion(acceptedSuggestion.content, session.startPosition)
-            // use the addedCharactersForEdit if it is EDIT suggestion type
-            const originalString = addedCharactersForEdit ? addedCharactersForEdit : acceptedSuggestion.content
-
-            codeDiffTracker.enqueue({
-                sessionId: session.codewhispererSessionId || '',
-                requestId: session.responseContext?.requestId || '',
-                fileUrl: session.document.uri,
-                languageId: session.language,
-                time: Date.now(),
-                originalString: originalString,
-                startPosition: session.startPosition,
-                endPosition: endPosition,
-                customizationArn: session.customizationArn,
-                completionType: getCompletionType(acceptedSuggestion),
-                triggerType: session.triggerType,
-                credentialStartUrl: session.credentialStartUrl,
-            })
-        }
-
-        const onLogInlineCompletionSessionResultsHandler = async (params: LogInlineCompletionSessionResultsParams) => {
-            const {
-                sessionId,
-                completionSessionResult,
-                firstCompletionDisplayLatency,
-                totalSessionDisplayTime,
-                typeaheadLength,
-                isInlineEdit,
-                addedDiagnostics,
-                removedDiagnostics,
-            } = params
-
-            const session = sessionManager.getSessionById(sessionId)
-            if (!session) {
-                logging.log(`ERROR: Session ID ${sessionId} was not found`)
-                return
-            }
-
-            if (session.state !== 'ACTIVE') {
-                logging.log(`ERROR: Trying to record trigger decision for not-active session ${sessionId}`)
-                return
-            }
-
-            const acceptedItemId = Object.keys(params.completionSessionResult).find(
-                k => params.completionSessionResult[k].accepted
-            )
-            const isAccepted = acceptedItemId ? true : false
-            const acceptedSuggestion = session.suggestions.find(s => s.itemId === acceptedItemId)
-            let addedCharactersForEditSuggestion = ''
-            let deletedCharactersForEditSuggestion = ''
-            if (acceptedSuggestion !== undefined) {
-                if (acceptedSuggestion) {
-                    codePercentageTracker.countSuccess(session.language)
-                    if (session.suggestionType === SuggestionType.EDIT && acceptedSuggestion.content) {
-                        // [acceptedSuggestion.insertText] will be undefined for NEP suggestion. Use [acceptedSuggestion.content] instead.
-                        // Since [acceptedSuggestion.content] is in the form of a diff, transform the content into addedCharacters and deletedCharacters.
-                        const addedAndDeletedChars = getAddedAndDeletedChars(acceptedSuggestion.content)
-                        if (addedAndDeletedChars) {
-                            addedCharactersForEditSuggestion = addedAndDeletedChars.addedCharacters
-                            deletedCharactersForEditSuggestion = addedAndDeletedChars.deletedCharacters
-
-                            codePercentageTracker.countAcceptedTokens(
-                                session.language,
-                                addedCharactersForEditSuggestion
-                            )
-                            codePercentageTracker.countTotalTokens(
-                                session.language,
-                                addedCharactersForEditSuggestion,
-                                true
-                            )
-                            enqueueCodeDiffEntry(session, acceptedSuggestion, addedCharactersForEditSuggestion)
-                        }
-                    } else if (acceptedSuggestion.insertText) {
-                        codePercentageTracker.countAcceptedTokens(session.language, acceptedSuggestion.insertText)
-                        codePercentageTracker.countTotalTokens(session.language, acceptedSuggestion.insertText, true)
-
-                        enqueueCodeDiffEntry(session, acceptedSuggestion)
-                    }
-                }
-            }
-
-            // Handle rejected edit predictions
-            if (isInlineEdit && !isAccepted) {
-                // Find all rejected suggestions in this session
-                const rejectedSuggestions = session.suggestions.filter(suggestion => {
-                    const result = completionSessionResult[suggestion.itemId]
-                    return result && result.seen && !result.accepted
-                })
-
-                // Record each rejected edit
-                for (const rejectedSuggestion of rejectedSuggestions) {
-                    if (rejectedSuggestion.content) {
-                        rejectedEditTracker.recordRejectedEdit({
-                            content: rejectedSuggestion.content,
-                            timestamp: Date.now(),
-                            documentUri: session.document.uri,
-                            position: session.startPosition,
-                        })
-
-                        logging.debug(
-                            `[EDIT_PREDICTION] Recorded rejected edit: ${rejectedSuggestion.content.substring(0, 20)}...`
-                        )
-                    }
-                }
-            }
-
-            session.setClientResultData(
-                completionSessionResult,
-                firstCompletionDisplayLatency,
-                totalSessionDisplayTime,
-                typeaheadLength
-            )
-
-            if (firstCompletionDisplayLatency) emitPerceivedLatencyTelemetry(telemetry, session)
-
-            // Always emit user trigger decision at session close
-            sessionManager.closeSession(session)
-            const streakLength = editsEnabled ? sessionManager.getAndUpdateStreakLength(isAccepted) : 0
-            await emitUserTriggerDecisionTelemetry(
-                telemetry,
-                telemetryService,
-                session,
-                timeSinceLastUserModification,
-                addedCharactersForEditSuggestion.length,
-                deletedCharactersForEditSuggestion.length,
-                addedDiagnostics,
-                removedDiagnostics,
-                streakLength
-            )
-        }
+        const apiController: CodeWhispererController = new CodeWhispererController()
+        const documentChangedListener = new DocumentChangedListener()
 
         const updateConfiguration = (updatedConfig: AmazonQWorkspaceConfig) => {
             logging.debug('Updating configuration of inline complete server.')
@@ -311,6 +132,34 @@ export const CodewhispererServerFactory =
             )
 
             await amazonQServiceManager.addDidChangeConfigurationListener(updateConfiguration)
+            apiController.init(
+                sessionManager,
+                logging,
+                lsp,
+                workspace,
+                amazonQServiceManager,
+                cursorTracker,
+                recentEditTracker,
+                codePercentageTracker,
+                userWrittenCodeTracker,
+                documentChangedListener,
+                telemetry,
+                telemetryService,
+                credentialsProvider,
+                rejectedEditTracker,
+                codeDiffTracker
+            )
+        }
+
+        const onInlineCompletionHandler = async (
+            params: InlineCompletionWithReferencesParams,
+            token: CancellationToken
+        ): Promise<InlineCompletionListWithReferences> => {
+            return apiController.onInlineCompletion(params, token)
+        }
+
+        const onLogInlineCompletionSessionResultsHandler = async (params: LogInlineCompletionSessionResultsParams) => {
+            return apiController.onLogInlineCompletionSessionResultsHandler(params)
         }
 
         lsp.extensions.onInlineCompletionWithReferences(onInlineCompletionHandler)
@@ -344,10 +193,7 @@ export const CodewhispererServerFactory =
             })
 
             // Record last user modification time for any document
-            if (lastUserModificationTime) {
-                timeSinceLastUserModification = new Date().getTime() - lastUserModificationTime
-            }
-            lastUserModificationTime = new Date().getTime()
+            documentChangedListener.onDocumentChanged(p)
 
             // Process document changes with RecentEditTracker.
             if (editsEnabled && recentEditTracker) {
