+++ server/aws-lsp-codewhisperer/src/language-server/inline-completion/codewhispererController.ts
@@ -1,11 +1,11 @@
 import {
     CancellationToken,
     CredentialsProvider,
-    InitializeParams,
     InlineCompletionListWithReferences,
     InlineCompletionWithReferencesParams,
     Logging,
     LogInlineCompletionSessionResultsParams,
+    Lsp,
     Telemetry,
     Workspace,
 } from '@aws/language-server-runtimes/server-interface'
@@ -17,58 +17,41 @@ import { DocumentChangedListener } from './documentChangedListener'
 import { SessionManager } from './session/sessionManager'
 import { RecentEditTracker } from './tracker/codeEditTracker'
 import { CursorTracker } from './tracker/cursorTracker'
-import { EditCompletionHandler } from './editCompletionHandler'
 import { InlineCompletionHandler } from './inlineCompletionHandler'
-import { LogInlineCompletionSessionResultsHandler } from './logInlineCompletionSessionResultsHandler'
 import { RejectedEditTracker } from './tracker/rejectedEditTracker'
 import { AcceptedInlineSuggestionEntry, CodeDiffTracker } from './codeDiffTracker'
+import { LogInlineCompletionSessionResultsHandler } from './logInlineCompletionSessionResultHandler'
 
 export class CodeWhispererController {
     // API handlers
-    editCompletionHandler: EditCompletionHandler | undefined
     private inlineCompletionHandler: InlineCompletionHandler | undefined
     private logInlineCompletionSessionResultsHandler: LogInlineCompletionSessionResultsHandler | undefined
 
-    constructor(
-        readonly sessionManager: SessionManager,
-        readonly editsSessionManager: SessionManager,
-        readonly logging: Logging,
-        readonly clientMetadata: InitializeParams,
-        readonly workspace: Workspace,
-        readonly qServiceManager: AmazonQBaseServiceManager,
-        readonly cursorTracker: CursorTracker,
-        readonly recentEditsTracker: RecentEditTracker,
-        readonly codePercentageTracker: CodePercentageTracker,
-        readonly userWrittenCodeTracker: UserWrittenCodeTracker | undefined,
-        readonly documentChangedListener: DocumentChangedListener,
-        readonly telemetry: Telemetry,
-        readonly telemetryService: TelemetryService,
-        readonly credentialsProvider: CredentialsProvider,
-        readonly rejectedEditTracker: RejectedEditTracker,
-        readonly codeDiffTracker: CodeDiffTracker<AcceptedInlineSuggestionEntry>
-    ) {
-        // Initialize api handler
-        this.editCompletionHandler = new EditCompletionHandler(
-            logging,
-            clientMetadata,
-            workspace,
-            qServiceManager,
-            this.editsSessionManager,
-            cursorTracker,
-            recentEditsTracker,
-            rejectedEditTracker,
-            documentChangedListener,
-            telemetry,
-            telemetryService,
-            credentialsProvider
-        )
+    constructor() {}
 
+    init(
+        sessionManager: SessionManager,
+        logging: Logging,
+        lsp: Lsp,
+        workspace: Workspace,
+        qServiceManager: AmazonQBaseServiceManager,
+        cursorTracker: CursorTracker,
+        recentEditsTracker: RecentEditTracker,
+        codePercentageTracker: CodePercentageTracker,
+        userWrittenCodeTracker: UserWrittenCodeTracker | undefined,
+        documentChangedListener: DocumentChangedListener,
+        telemetry: Telemetry,
+        telemetryService: TelemetryService,
+        credentialsProvider: CredentialsProvider,
+        rejectedEditTracker: RejectedEditTracker,
+        codeDiffTracker: CodeDiffTracker<AcceptedInlineSuggestionEntry>
+    ) {
         this.inlineCompletionHandler = new InlineCompletionHandler(
             logging,
-            clientMetadata,
+            lsp,
             workspace,
             qServiceManager,
-            this.sessionManager,
+            sessionManager,
             cursorTracker,
             recentEditsTracker,
             codePercentageTracker,
@@ -76,20 +59,8 @@ export class CodeWhispererController {
             documentChangedListener,
             telemetry,
             telemetryService,
-            credentialsProvider
-        )
-
-        this.logInlineCompletionSessionResultsHandler = new LogInlineCompletionSessionResultsHandler(
-            logging,
-            clientMetadata,
-            this.sessionManager,
-            this.editsSessionManager,
-            codePercentageTracker,
-            codeDiffTracker,
-            rejectedEditTracker,
-            documentChangedListener,
-            telemetry,
-            telemetryService
+            credentialsProvider,
+            rejectedEditTracker
         )
     }
 
@@ -98,26 +69,15 @@ export class CodeWhispererController {
         token: CancellationToken
     ): Promise<InlineCompletionListWithReferences> {
         if (!this.inlineCompletionHandler) {
-            throw new Error('')
+            throw new Error('inlineCompletionHandler is not initialized yet')
         }
 
         return this.inlineCompletionHandler.onInlineCompletion(params, token)
     }
 
-    async onEditCompletion(
-        params: InlineCompletionWithReferencesParams,
-        token: CancellationToken
-    ): Promise<InlineCompletionListWithReferences> {
-        if (!this.editCompletionHandler) {
-            throw new Error('')
-        }
-
-        return this.editCompletionHandler.onEditCompletion(params, token)
-    }
-
     async onLogInlineCompletionSessionResultsHandler(params: LogInlineCompletionSessionResultsParams) {
         if (!this.logInlineCompletionSessionResultsHandler) {
-            throw new Error('')
+            throw new Error('logInlineCompletionSessionResultsHandler is not intialized yet')
         }
 
         await this.logInlineCompletionSessionResultsHandler.onLogInlineCompletionSessionResultsHandler(params)
