+++ server/aws-lsp-codewhisperer/src/shared/codeWhispererService.ts
@@ -257,6 +257,7 @@ export class CodeWhispererServiceToken extends CodeWhispererServiceBase {
         const response = await this.client.generateCompletions(this.withProfileArn(request)).promise()
         this.logging.info(
             `GenerateCompletion response: 
+    "version": "refactor",
     "endpoint": ${this.codeWhispererEndpoint},
     "requestId": ${response.$response.requestId},
     "responseCompletionCount": ${response.completions?.length ?? 0},
